yongsik:[Salt-Client] change to call using the information got to the server. #85
