+++ .vscode/launch.json
@@ -18,9 +18,9 @@
         "-d",
         "--port=443",
         "--cert=${workspaceFolder}/backend/cmd/cmp/scmp_self_signed.pem",
-        "-u=salt:http://61.250.122.43:8000/",
+        "-u=salt:http://61.250.122.43:8000/run",
         "-u=128T-Oncue:https://211.189.153.40/api/v1/graphql",
-        "-k=salt:9853e2c223032e49cc7328de5969c6b3d0eadbb6",
+        "-k=salt:d579ab9d2f188d8d9beb8a65aabc80b5db2506b2e5989e526b9616d427e55201",
         "-k=128T-Oncue:eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJuYW1lIjoiYWRtaW4iLCJzY29wZXMiOlsiY29uZmlndXJlIl0sImlhdCI6MTU3NDAzNjAxM30.beW77E48pOhNKHNuFB6q343hg5XGbV7HzUcI-8LaV4RJTOiOALpA4NPWF4nDmFcveVuGuH8rETnLNbeXFa2zmssbCg_TfaAt49yyq_Meg43TYXKOYl5pl7mJ1k-nte0zW4fX2J6brRE5ouvU3e2ssvRVmvsnuqWxdQ1Doqm4e2zLaVIblMsb80DpSr7AZfbF5Ld11KwUPIxAsU7LIN_tfKfy4-LPqwN78eD6fGcb_u6m2sm5jVZihuwKmzhIVEVIHtnGbr2oSrAHLrpeokLCGctXF1b7pCt7pCWIeBqyr7EjjOx90d8LoH49idMPLymDR_NPi-8Cs0GXszJtW82DuA",
         "-c=./cmp-canned/",
         "--protoboards-path=./cmp-protoboards/",
