+++ frontend/src/agent_admin/containers/AgentControl.tsx
@@ -10,13 +10,15 @@ import AgentControlTable from 'src/agent_admin/components/AgentControlTable'
 import AgentControlConsole from 'src/agent_admin/components/AgentControlConsole'
 
 // APIs
+import {getMinionKeyListAllAsync} from 'src/agent_admin/apis'
+
+// SaltStack
 import {
-  getMinionKeyListAllAsync,
   runLocalServiceStartTelegraf,
   runLocalServiceStopTelegraf,
   runLocalCpGetDirTelegraf,
   runLocalPkgInstallTelegraf,
-} from 'src/agent_admin/apis'
+} from 'src/shared/apis/saltStack'
 
 // Notification
 import {notify as notifyAction} from 'src/shared/actions/notifications'
@@ -65,7 +67,11 @@ export class AgentControl extends PureComponent<Props, State> {
   }
 
   getWheelKeyListAll = async () => {
-    const hostListObject = await getMinionKeyListAllAsync()
+    const {saltMasterUrl, saltMasterToken} = this.props
+    const hostListObject = await getMinionKeyListAllAsync(
+      saltMasterUrl,
+      saltMasterToken
+    )
 
     this.setState({
       Minions: _.values(hostListObject),
@@ -75,11 +81,11 @@ export class AgentControl extends PureComponent<Props, State> {
   }
 
   public async componentWillMount() {
-    const {notify, saltMasterToken} = this.props
-    if (saltMasterToken !== null && saltMasterToken !== '') {
+    const {notify} = this.props
+    try {
       this.getWheelKeyListAll()
       this.setState({controlPageStatus: RemoteDataState.Loading})
-    } else {
+    } catch (e) {
       this.setState({controlPageStatus: RemoteDataState.Done})
       notify(notifyAgentConnectFailed('Token is not valid.'))
     }
@@ -133,8 +139,11 @@ export class AgentControl extends PureComponent<Props, State> {
   }
 
   public onClickActionCall = (host: string, isRunning: boolean) => () => {
+    const {saltMasterUrl, saltMasterToken} = this.props
     if (isRunning === false) {
       const getLocalServiceStartTelegrafPromise = runLocalServiceStartTelegraf(
+        saltMasterUrl,
+        saltMasterToken,
         host
       )
 
@@ -153,6 +162,8 @@ export class AgentControl extends PureComponent<Props, State> {
       )
     } else {
       const getLocalServiceStopTelegrafPromise = runLocalServiceStopTelegraf(
+        saltMasterUrl,
+        saltMasterToken,
         host
       )
       this.setState({controlPageStatus: RemoteDataState.Loading})
@@ -169,6 +180,7 @@ export class AgentControl extends PureComponent<Props, State> {
   }
 
   public onClickRunCall = () => {
+    const {saltMasterUrl, saltMasterToken} = this.props
     const {Minions} = this.state
     const host = Minions.filter(m => m.isCheck === true).map(
       checkData => checkData.host
@@ -176,6 +188,8 @@ export class AgentControl extends PureComponent<Props, State> {
 
     this.setState({controlPageStatus: RemoteDataState.Loading})
     const getLocalServiceStartTelegrafPromise = runLocalServiceStartTelegraf(
+      saltMasterUrl,
+      saltMasterToken,
       host
     )
 
@@ -191,13 +205,18 @@ export class AgentControl extends PureComponent<Props, State> {
   }
 
   public onClickStopCall = () => {
+    const {saltMasterUrl, saltMasterToken} = this.props
     const {Minions} = this.state
     const host = Minions.filter(m => m.isCheck === true).map(
       checkData => checkData.host
     )
 
     this.setState({controlPageStatus: RemoteDataState.Loading})
-    const getLocalServiceStopTelegrafPromise = runLocalServiceStopTelegraf(host)
+    const getLocalServiceStopTelegrafPromise = runLocalServiceStopTelegraf(
+      saltMasterUrl,
+      saltMasterToken,
+      host
+    )
 
     getLocalServiceStopTelegrafPromise.then(pLocalServiceStopTelegrafData => {
       this.setState({
@@ -212,6 +231,7 @@ export class AgentControl extends PureComponent<Props, State> {
   }
 
   public onClickInstallCall = () => {
+    const {saltMasterUrl, saltMasterToken} = this.props
     const {Minions} = this.state
 
     const host = Minions.filter(m => m.isCheck === true).map(
@@ -220,7 +240,11 @@ export class AgentControl extends PureComponent<Props, State> {
 
     this.setState({controlPageStatus: RemoteDataState.Loading})
 
-    const getLocalCpGetDirTelegrafPromise = runLocalCpGetDirTelegraf(host)
+    const getLocalCpGetDirTelegrafPromise = runLocalCpGetDirTelegraf(
+      saltMasterUrl,
+      saltMasterToken,
+      host
+    )
 
     getLocalCpGetDirTelegrafPromise.then(pLocalCpGetDirTelegrafData => {
       this.setState({
@@ -230,7 +254,11 @@ export class AgentControl extends PureComponent<Props, State> {
           yaml.dump(pLocalCpGetDirTelegrafData.data.return[0]),
       })
 
-      const getLocalPkgInstallTelegrafPromise = runLocalPkgInstallTelegraf(host)
+      const getLocalPkgInstallTelegrafPromise = runLocalPkgInstallTelegraf(
+        saltMasterUrl,
+        saltMasterToken,
+        host
+      )
 
       getLocalPkgInstallTelegrafPromise.then(pLocalPkgInstallTelegrafData => {
         this.setState({
